Revert moving FatturaElettronicaHeaderType
